gaia CLI fixes and dep fixes
