+++ cmd/gaia/node.go
@@ -1,12 +1,9 @@
 package main
 
 import (
-	"os"
-
 	"github.com/spf13/cobra"
 
 	abci "github.com/tendermint/abci/types"
-	"github.com/tendermint/tmlibs/cli"
 
 	sdk "github.com/cosmos/cosmos-sdk"
 	"github.com/cosmos/cosmos-sdk/modules/auth"
@@ -21,47 +18,17 @@ import (
 	"github.com/cosmos/cosmos-sdk/state"
 
 	"github.com/cosmos/gaia/modules/stake"
-	"github.com/cosmos/gaia/version"
 )
 
 // nodeCmd is the entry point for this binary
 var nodeCmd = &cobra.Command{
 	Use:   "node",
 	Short: "The Cosmos Network delegation-game blockchain test",
-	PreRun: func(cmd *cobra.Command, args []string) {
-		_ = cli.PrepareMainCmd(cmd, "GA", os.ExpandEnv("$HOME/.cosmos-gaia"))
-	},
-	Run: func(cmd *cobra.Command, args []string) { cmd.Help() },
+	Run:   func(cmd *cobra.Command, args []string) { cmd.Help() },
 }
 
 func prepareNodeCommands() {
-	nodeCmd.AddCommand(
-		basecmd.GetInitCmd("fermion", []string{"stake/allowed_bond_denom/fermion"}),
-		basecmd.GetTickStartCmd(sdk.TickerFunc(tickFn)),
-		basecmd.UnsafeResetAllCmd,
-		version.VersionCmd,
-	)
-}
 
-// Tick - Called every block even if no transaction,
-// process all queues, validator rewards, and calculate the validator set difference
-func tickFn(ctx sdk.Context, store state.SimpleDB) (diffVal []*abci.Validator, err error) {
-	// First need to prefix the store, at this point it's a global store
-	store = stack.PrefixedStore(stake.Name(), store)
-
-	// Determine the validator set changes
-	candidates := stake.LoadCandidates(store)
-	startVal := candidates.GetValidators(store)
-	changed := candidates.UpdateVotingPower(store)
-	if !changed {
-		return
-	}
-	newVal := candidates.GetValidators(store)
-	diffVal = stake.ValidatorsDiff(startVal, newVal, store)
-	return
-}
-
-func cmdNode(cmd *cobra.Command, args []string) error {
 	// require all fees in strings - change this in your app!
 	basecmd.Handler = stack.New(
 		base.Logger{},
@@ -84,7 +51,28 @@ func cmdNode(cmd *cobra.Command, args []string) error {
 			stake.NewHandler(),
 		)
 
-	basecmd.SetUpRoot(cmd)
+	nodeCmd.AddCommand(
+		basecmd.GetInitCmd("fermion", []string{"stake/allowed_bond_denom/fermion"}),
+		basecmd.GetTickStartCmd(sdk.TickerFunc(tickFn)),
+		basecmd.UnsafeResetAllCmd,
+	)
+	//basecmd.SetUpRoot(nodeCmd)
+}
 
-	return nil
+// Tick - Called every block even if no transaction,
+// process all queues, validator rewards, and calculate the validator set difference
+func tickFn(ctx sdk.Context, store state.SimpleDB) (diffVal []*abci.Validator, err error) {
+	// First need to prefix the store, at this point it's a global store
+	store = stack.PrefixedStore(stake.Name(), store)
+
+	// Determine the validator set changes
+	candidates := stake.LoadCandidates(store)
+	startVal := candidates.GetValidators(store)
+	changed := candidates.UpdateVotingPower(store)
+	if !changed {
+		return
+	}
+	newVal := candidates.GetValidators(store)
+	diffVal = stake.ValidatorsDiff(startVal, newVal, store)
+	return
 }
