+++ cmd/gaia/rest.go
@@ -4,7 +4,6 @@ import (
 	"fmt"
 	"log"
 	"net/http"
-	"os"
 
 	"github.com/gorilla/mux"
 	"github.com/spf13/cobra"
@@ -16,47 +15,30 @@ import (
 	noncerest "github.com/cosmos/cosmos-sdk/modules/nonce/rest"
 	rolerest "github.com/cosmos/cosmos-sdk/modules/roles/rest"
 
-	"github.com/tendermint/tmlibs/cli"
-
 	stakerest "github.com/cosmos/gaia/modules/stake/rest"
 )
 
 const defaultAlgo = "ed25519"
 
 var (
-	serverCmd = &cobra.Command{
-		Use:   "server",
+	restServerCmd = &cobra.Command{
+		Use:   "rest-server",
 		Short: "REST client for gaia commands",
 		Long:  `Gaiaserver presents  a nice (not raw hex) interface to the gaia blockchain structure.`,
-		PreRun: func(cmd *cobra.Command, args []string) {
-			// this should share the dir with gaiacli, so you can use the cli and
-			// the api interchangeably
-			_ = cli.PrepareMainCmd(cmd, "GA", os.ExpandEnv("$HOME/.gaiacli"))
-		},
-
-		Run: func(cmd *cobra.Command, args []string) { cmd.Help() },
-	}
-
-	serveCmd = &cobra.Command{
-		Use:   "serve",
-		Short: "Serve the REST client",
 		RunE: func(cmd *cobra.Command, args []string) error {
-			return cmdServe(cmd, args)
+			return cmdRestServer(cmd, args)
 		},
 	}
 
 	flagPort = "port"
 )
 
-func prepareServerCommands() {
-	serverCmd.AddCommand(commands.InitCmd)
-	serverCmd.AddCommand(commands.VersionCmd)
-	serverCmd.AddCommand(serveCmd)
-	commands.AddBasicFlags(serveCmd)
-	serveCmd.PersistentFlags().IntP(flagPort, "p", 8998, "port to run the server on")
+func prepareRestServerCommands() {
+	commands.AddBasicFlags(restServerCmd)
+	restServerCmd.PersistentFlags().IntP(flagPort, "p", 8998, "port to run the server on")
 }
 
-func cmdServe(cmd *cobra.Command, args []string) error {
+func cmdRestServer(cmd *cobra.Command, args []string) error {
 	router := mux.NewRouter()
 
 	routeRegistrars := []func(*mux.Router) error{
