+++ glide.lock
@@ -10,7 +10,7 @@ imports:
 - name: github.com/BurntSushi/toml
   version: a368813c5e648fee92e5f6c30e3944ff9d5e8895
 - name: github.com/cosmos/cosmos-sdk
-  version: 0e14fd69102c6ecf49ac62c945030faae98a42eb
+  version: c734af9f1019ffa35280d8d5f5494a4dc9c5e2ae
   subpackages:
   - app
   - client
